requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pandas>=1.5.0
plotly>=5.15.0
python-dotenv>=1.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
pillow>=10.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0